+++ exporter/opentelemetry-exporter-prometheus-remote-write/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
 ## Unreleased
-- Setup for Prometheus Remote Write Exporter
+- Prometheus Remote Write Exporter Setup
   ((#180)[https://github.com/open-telemetry/opentelemetry-python-contrib/pull/180])
+- Add Exporter constructor validation methods
+  ((#206)[https://github.com/open-telemetry/opentelemetry-python-contrib/pull/206])